feat: refactored and removed unnecessary data and fixed the theme switcher component
