+++ portfolio-project/src/components/Hero/TypingHeader.tsx
