--- portfolio-project/src/components/utils/educationData.tsx
@@ -1,20 +1,0 @@
-export const educationData = [
-    {
-        edu: "Full-stack Engineering with Agile Methodologies",
-        school: "Code Chrysalis",
-        date: "06/2024 – 09/2024",
-        location: "Tokyo, Japan",
-    },
-    {
-        edu: "Japanese",
-        school: "Akamonkai Language School",
-        date: "10/2023 – 04/2024",
-        location: "Tokyo, Japan",
-    },
-    {
-        edu: "Bachelor of Science - International Business",
-        school: "Temple University",
-        date: "09/2014 – 05/2018",
-        location: "Tokyo, Japan",
-    },
-];
