--- portfolio-project/src/components/utils/professionalExperienceData.tsx
@@ -1,36 +1,0 @@
-export const professionalExperienceData = [
-    {
-        role: "Solutions Consultant",
-        company: "ESG Global",
-        date: "01/2025 - present",
-        roleKeyPoints: [
-            "Managing multiple strict test deadlines ensuring effective client adoption and onboarding into new energy markets",
-            "Responsible for ensuring proper file transmission, data flow, and database integrity to ensure effective communication between clients and utilities.",
-            "Owning and monitoring market requirements to ensure business rules are up-to-date.",
-        ],
-    },
-    {
-        role: "Applications Consultant",
-        company: "ESG Global",
-        date: "10/2021 - 07/2023",
-        roleKeyPoints: [
-            "Acted as the primary point of contact for technical issues developing technical proficiency with all proprietary solutions within the Japan team, establishing and maintaining strong relationships at all levels within our organization.",
-            "Performed regular API testing as part of our operations processes, managed server health, and conducted database querying and analysis utilizing SQL to ensure application stability.",
-            "Bridged communications among cross-functional teams within an Agile framework facilitating seamless collaboration and information sharing.",
-            "Headed the User Acceptance Testing (UAT) initiative for new releases in the Japanese market, ensuring smooth adoption by clients.",
-            "Led the testing process for multiple projects across clients, providing the PMO team with valuable system functionality and procedural insights.",
-            "Collaborated with PMO teams for backlog grooming and JIRA prioritization, contributing to efficient project management.",
-        ],
-    },
-    {
-        role: "Client Operations Analyst",
-        company: "ESG Global",
-        date: "08/2018 - 09/2021",
-        roleKeyPoints: [
-            "Established and maintained strong relationships with Japanese clients as primary point of contact, establishing and maintaining strong client relationships",
-            "Mentored and trained team members as team lead on software issue resolutions and responses",
-            "Advised clients on best practices for utilizing our platform including developing and sharing instructional documents",
-            "Collaborated with PMO and SD teams on early requirements gathering representing the client's needs.",
-        ],
-    },
-]
