--- portfolio-project/src/components/utils/projectData.tsx
@@ -1,108 +1,0 @@
-export const projectData = [
-    {
-        title: "Hato",
-        desc: "A social media app designed around flashcards",
-        role: "Full-stack Engineer",
-        keyPoints: [
-            "Frontend built utilizing Typescript, React, and CSS",
-            "Backend built utilizing Node.js",
-            "Developed as a fullstack developer among a group of 4 in an agile environment",
-        ],
-        tech: [
-            "Express.js", 
-            "Node.js", 
-            "React.js", 
-            "Knex.js", 
-            "PostgreSQL", 
-            "JavaScript", 
-            "TypeScript"
-        ],
-        image: "/project-captures/Hato.JPG",
-        link: "https://github.com/hato-app"
-    },
-    {
-        title: "Dokushojo",
-        desc: "A Language Learning app focusing on listening comprehension",
-        role: "Full-stack/Project Manager",
-        keyPoints: [
-            "Frontend built utilizing TypeScript, React.js, Vite and Bootstrap.",
-            "Backend built utilizing Node.js, Knex.js and PostgreSQL, and Express.js.",
-            "Managed project backlog and maintained project schedule in a team of 4",
-        ],
-        tech: [
-            "Express.js", 
-            "Node.js", 
-            "React.js", 
-            "Knex.js", 
-            "Bootstrap CSS", 
-            "PostgreSQL", 
-            "Vite", 
-            "TypeScript"
-        ],
-        image: "/project-captures/DSJ.JPG",
-        link: "https://github.com/Planet-DokuShojo/"
-    },
-    {
-        title: "VitaeAggregate",
-        desc: "A community app designed around building ATS friendly resumes",     
-        role: "Full-stack/Project Manager",
-        keyPoints: [
-            "Frontend built utilizing TypeScript, Svelte, Tailwind, Firebase for authentication, and Puppeteer.",
-            "Backend built utilizing Python, Django REST, and PostgreSQL.",
-            "Version control with Git and containerization with Docker",
-            "Managed project schedule, project backlog, and led daily standups in an agile development environment.",
-        ],
-        tech: [
-            "Cypress", 
-            "Node.js", 
-            "Svelte.js", 
-            "Django", 
-            "PostgreSQL", 
-            "Python", 
-            "TypeScript", 
-            "Pytest", 
-            "Docker", 
-            "Vercel"
-        ],
-        image: "/project-captures/VitaeAggregate.JPG",
-        link: "https://github.com/vitaeaggregate"
-    },
-    {
-        title: "Landing Site",
-        desc: "A responsive, mobile-first homepage built for a local musical act in the Tokyo area. ",
-        role: "Front-end/Project Lead",
-        keyPoints: [
-            "Built using Next.js, Typescript, and Tailwind",
-            "Deployed utilizing Vercel",
-            "Animations utilizing framer-motion",
-        ],
-        tech: [
-            "Next.js", 
-            "Node.js", 
-            "Tailwind", 
-            "TypeScript", 
-            "Vercel"
-        ],
-        image: "/project-captures/Bandpage.JPG",
-        link: "https://github.com/sentient-band-site"
-    },
-    {
-        title: "Portfolio Site",
-        desc: "A portfolio showcasing my work",
-        role: "Front-end/Project Lead",
-        keyPoints: [
-            "Built using Next.js, Typescript, and Tailwind",
-            "Deployed utilizing Vercel",
-        ],
-        tech: [
-            "Next.js", 
-            "Node.js", 
-            "Tailwind", 
-            "TypeScript", 
-            "Vercel"
-        ],
-        image: "/project-captures/Portfolio.JPG",
-        link: "https://github.com/Bcook2289/portfolio"
-    },
-
-];
